requirements.txt
pandas~=2.2.2
dash~=2.17.1
plotly~=5.22.0
google-cloud-storage
